requirements.txt
flask
flask-socketio
flask-login
werkzeug
google-genai
pillow
google-cloud-vision
requests
